+++ react/AddressContainer.js
@@ -32,7 +32,7 @@ class AddressContainer extends Component {
     }
   }
 
-  handleAddressChange = (changedAddressFields) => {
+  handleAddressChange = changedAddressFields => {
     const {
       cors,
       accountName,
@@ -67,15 +67,16 @@ class AddressContainer extends Component {
       !changedAddressFields.postalCode.geolocationAutoCompleted
     ) {
       const postalCodeField =
-        rules.fields &&
-        rules.fields.find((field) => field.name === 'postalCode')
+        rules.fields && rules.fields.find(field => field.name === 'postalCode')
       const diffFromPrev =
         address.postalCode.value !== validatedAddress.postalCode.value
       const isValidPostalCode = validatedAddress.postalCode.valid === true
+      const isValidCountryCode = validatedAddress.country.valid === true
       const shouldAutoComplete =
         rules.postalCodeFrom === POSTAL_CODE &&
         diffFromPrev &&
         isValidPostalCode &&
+        isValidCountryCode &&
         postalCodeField &&
         postalCodeField.postalCodeAPI